+++ eras/dijkstra/impl/testlib/Test/Cardano/Ledger/Dijkstra/Arbitrary.hs
@@ -124,6 +124,15 @@ instance Arbitrary (UpgradeDijkstraPParams Identity DijkstraEra) where
       <*> arbitrary
       <*> arbitrary
       <*> genValidCostModel PlutusV4
+      <*> arbitrary
+      <*> arbitrary
+      <*> arbitrary
+      <*> arbitrary
+      <*> arbitrary
+      <*> arbitrary
+      <*> arbitrary
+      <*> arbitrary
+      <*> arbitrary
 
 instance Arbitrary DijkstraGenesis where
   arbitrary = genericArbitraryU