+++ libs/cardano-ledger-core/testlib/Test/Cardano/Ledger/TreeDiff.hs
@@ -206,6 +206,8 @@ instance (Typeable r, ToExpr a) => ToExpr (Mismatch r a) where
         , ("expected", toExpr expected)
         ]
 
+instance ToExpr SlotInterval
+
 instance ToExpr EpochInterval
 
 -- AuxiliaryData